fix: clear html and added links nav
