+++ movie-app/src/components/Navigation/Navigation.css
@@ -15,6 +15,8 @@
   display: flex;
   flex-direction: column;
   align-items: end;
+  text-align: center;
+  /* background-color: #fff; */
   -webkit-user-select: none;
   user-select: none;
 }
