+++ movie-app/src/components/PageNotFound/PageNotFound.css
@@ -42,3 +42,14 @@
   padding: 0;
   margin: 0;
 }
+
+.not-found__link {
+  text-decoration: none;
+  cursor: pointer;
+  font-size: 12px;
+  line-height: 15px;
+  /* identical to box height */
+  text-align: center;
+  color: #000;
+  margin-bottom: 30px;
+}
