+++ movie-app/src/components/SavedMovies/SavedMovies.jsx
@@ -1,12 +1,13 @@
 import '../SavedMovies/SavedMovies.css'
 import MoviesCardList from '../MoviesCardList/MoviesCardList';
 import MoviesCard from '../MoviesCard/MoviesCard';
+import { moviesData } from '../../utils/movies'
 
 function SavedMovies() {
   return (
     <>
-      <MoviesCard />
       <MoviesCardList />
+       <MoviesCard movie={moviesData}/>
     </>
   );
 }
